// Desktop/final-project-yuxuan-weli-main 2/client/src/types/entityTypes.ts
interface UserResponseType {
    success: boolean;
    message?: string;
    token?: string;
    user?: {
        username: string;
        email: string;
        createdAt: string;
    };
}